+++ server/svix-server/src/core/message_app.rs
@@ -17,7 +17,9 @@ use crate::{
             OrganizationId,
         },
     },
+    ctx,
     db::models::{application, endpoint, message},
+    err_validation,
     error::{Error, Result},
 };
 
@@ -40,9 +42,7 @@ impl CreateMessageApp {
         db: &DatabaseTransaction,
         app: application::Model,
     ) -> Result<CreateMessageApp> {
-        let endpoints = endpoint::Entity::secure_find(app.id.clone())
-            .all(db)
-            .await?
+        let endpoints = ctx!(endpoint::Entity::secure_find(app.id.clone()).all(db).await)?
             .into_iter()
             .map(TryInto::try_into)
             .collect::<Result<Vec<_>>>()?;
@@ -55,9 +55,7 @@ impl CreateMessageApp {
                 .rate_limit
                 .map(|v| v.try_into())
                 .transpose()
-                .map_err(|_| {
-                    Error::Validation("Application rate limit out of bounds".to_owned())
-                })?,
+                .map_err(|_| err_validation!("Application rate limit out of bounds"))?,
             endpoints,
             deleted: app.deleted,
         })
@@ -82,14 +80,15 @@ impl CreateMessageApp {
         }
 
         // Then check PostgreSQL
-        let db = pg.begin().await?;
+        let db = ctx!(pg.begin().await)?;
         // Fetch the [`application::Model`] either given or from the ID
         let app = if let Some(app) = app {
             app
-        } else if let Some(app) = application::Entity::secure_find_by_id(org_id, app_id)
-            .one(&db)
-            .await?
-        {
+        } else if let Some(app) = ctx!(
+            application::Entity::secure_find_by_id(org_id, app_id)
+                .one(&db)
+                .await
+        )? {
             app
         } else {
             return Ok(None);
@@ -189,7 +188,7 @@ impl TryFrom<endpoint::Model> for CreateMessageEndpoint {
                 .rate_limit
                 .map(|v| v.try_into())
                 .transpose()
-                .map_err(|_| Error::Validation("Endpoint rate limit out of bounds".to_owned()))?,
+                .map_err(|_| err_validation!("Endpoint rate limit out of bounds"))?,
             first_failure_at: m.first_failure_at,
             headers: m.headers,
             disabled: m.disabled,