+++ server/svix-server/src/queue/memory.rs
@@ -4,7 +4,7 @@ use axum::async_trait;
 use chrono::Utc;
 use tokio::{sync::mpsc, time::sleep};
 
-use crate::error::{Error, Result};
+use crate::{err_queue, error::Result};
 
 use super::{
     QueueTask, TaskQueueConsumer, TaskQueueDelivery, TaskQueueProducer, TaskQueueReceive,
@@ -66,7 +66,7 @@ impl TaskQueueReceive for MemoryQueueConsumer {
                     tracing::trace!("MemoryQueue: event recv <");
                     Ok(vec![delivery])
                 } else {
-                    Err(Error::Queue("Failed to fetch from queue".to_owned()))
+                    Err(err_queue!("Failed to fetch from queue"))
                 }
             }
         }