+++ server/svix-server/src/v1/endpoints/message.rs
@@ -5,13 +5,14 @@ use crate::{
     cache::Cache,
     core::{
         message_app::CreateMessageApp,
-        security::{AuthenticatedApplication, AuthenticatedOrganizationWithApplication},
+        permissions,
         types::{
             ApplicationIdOrUid, EventChannel, EventChannelSet, EventTypeName, EventTypeNameSet,
             MessageAttemptTriggerType, MessageId, MessageIdOrUid, MessageUid,
         },
     },
-    error::{Error, HttpError, Result},
+    ctx, err_generic,
+    error::{HttpError, Result},
     queue::{MessageTaskBatch, TaskQueueProducer},
     v1::utils::{
         apply_pagination, iterator_from_before_or_after, validation_error, ListResponse,
@@ -172,10 +173,7 @@ async fn list_messages(
         after,
     }): ValidatedQuery<ListMessagesQueryParams>,
     list_filter: MessageListFetchOptions,
-    AuthenticatedApplication {
-        permissions: _,
-        app,
-    }: AuthenticatedApplication,
+    permissions::Application { app }: permissions::Application,
 ) -> Result<Json<ListResponse<MessageOut>>> {
     let PaginationLimit(limit) = pagination.limit;
 
@@ -202,9 +200,13 @@ async fn list_messages(
     };
 
     let out = if is_prev {
-        query.all(db).await?.into_iter().rev().map(into).collect()
+        ctx!(query.all(db).await)?
+            .into_iter()
+            .rev()
+            .map(into)
+            .collect()
     } else {
-        query.all(db).await?.into_iter().map(into).collect()
+        ctx!(query.all(db).await)?.into_iter().map(into).collect()
     };
 
     Ok(Json(MessageOut::list_response(out, limit as usize, false)))
@@ -224,26 +226,26 @@ async fn create_message(
         CreateMessageQueryParams,
     >,
     ValidatedJson(data): ValidatedJson<MessageIn>,
-    AuthenticatedOrganizationWithApplication { permissions, app }: AuthenticatedOrganizationWithApplication,
+    permissions::OrganizationWithApplication { app }: permissions::OrganizationWithApplication,
 ) -> Result<(StatusCode, Json<MessageOut>)> {
     let create_message_app = CreateMessageApp::layered_fetch(
         cache,
         db,
         Some(app.clone()),
         app.id.clone(),
-        app.org_id,
+        app.org_id.clone(),
         std::time::Duration::from_secs(30),
     )
     .await?
     // Should never happen since you're giving it an existing Application, but just in case
-    .ok_or_else(|| Error::Generic(format!("Application doesn't exist: {}", app.id)))?;
+    .ok_or_else(|| err_generic!("Application doesn't exist: {}", app.id))?;
 
     let msg = message::ActiveModel {
         app_id: Set(app.id.clone()),
-        org_id: Set(permissions.org_id),
+        org_id: Set(app.org_id),
         ..data.into()
     };
-    let msg = msg.insert(db).await?;
+    let msg = ctx!(msg.insert(db).await)?;
 
     let trigger_type = MessageAttemptTriggerType::Scheduled;
     if !create_message_app
@@ -280,15 +282,14 @@ async fn get_message(
     Extension(ref db): Extension<DatabaseConnection>,
     Path((_app_id, msg_id)): Path<(ApplicationIdOrUid, MessageIdOrUid)>,
     ValidatedQuery(GetMessageQueryParams { with_content }): ValidatedQuery<GetMessageQueryParams>,
-    AuthenticatedApplication {
-        permissions: _,
-        app,
-    }: AuthenticatedApplication,
+    permissions::Application { app }: permissions::Application,
 ) -> Result<Json<MessageOut>> {
-    let msg = message::Entity::secure_find_by_id_or_uid(app.id, msg_id)
-        .one(db)
-        .await?
-        .ok_or_else(|| HttpError::not_found(None, None))?;
+    let msg = ctx!(
+        message::Entity::secure_find_by_id_or_uid(app.id, msg_id)
+            .one(db)
+            .await
+    )?
+    .ok_or_else(|| HttpError::not_found(None, None))?;
     let msg_out = if with_content {
         msg.into()
     } else {