+++ server/svix-server/src/worker.rs
@@ -19,10 +19,11 @@ use crate::core::{
     },
 };
 use crate::db::models::{endpoint, eventtype, message, messageattempt, messagedestination};
-use crate::error::{Error, Result};
+use crate::error::Result;
 use crate::queue::{
     MessageTask, MessageTaskBatch, QueueTask, TaskQueueConsumer, TaskQueueProducer,
 };
+use crate::{ctx, err_generic};
 use chrono::Utc;
 
 use futures::future;
@@ -101,10 +102,7 @@ async fn process_success_cache(
 ) -> Result<()> {
     let key = FailureCacheKey::new(app_id, endp_id);
 
-    cache
-        .delete(&key)
-        .await
-        .map_err(|e| Error::from(e.to_string()))?;
+    cache.delete(&key).await.map_err(|e| err_generic!(e))?;
 
     Ok(())
 }
@@ -135,7 +133,7 @@ async fn process_failure_cache(
     if let Some(FailureCacheValue { first_failure_at }) = cache
         .get::<FailureCacheValue>(&key)
         .await
-        .map_err(|e| Error::from(e.to_string()))?
+        .map_err(|e| err_generic!(e))?
     {
         if now - first_failure_at
             > chrono::Duration::from_std(disable_in).expect("Given `disable_in` is too large")
@@ -158,7 +156,7 @@ async fn process_failure_cache(
                 disable_in * 2,
             )
             .await
-            .map_err(|e| Error::from(e.to_string()))?;
+            .map_err(|e| err_generic!(e))?;
 
         Ok(None)
     }
@@ -330,16 +328,13 @@ async fn dispatch(
         .send()
         .await;
 
-    let msg_dest = messagedestination::Entity::secure_find_by_msg(msg_task.msg_id.clone())
-        .filter(messagedestination::Column::EndpId.eq(endp.id.clone()))
-        .one(db)
-        .await?
-        .ok_or_else(|| {
-            Error::Generic(format!(
-                "Msg dest not found {} {}",
-                msg_task.msg_id, endp.id
-            ))
-        })?;
+    let msg_dest = ctx!(
+        messagedestination::Entity::secure_find_by_msg(msg_task.msg_id.clone())
+            .filter(messagedestination::Column::EndpId.eq(endp.id.clone()))
+            .one(db)
+            .await
+    )?
+    .ok_or_else(|| err_generic!("Msg dest not found {} {}", msg_task.msg_id, endp.id))?;
 
     if (msg_dest.status != MessageStatus::Pending && msg_dest.status != MessageStatus::Sending)
         && (msg_task.trigger_type != MessageAttemptTriggerType::Manual)
@@ -375,23 +370,35 @@ async fn dispatch(
             };
             let http_error = res.error_for_status_ref().err();
 
-            let bytes = res
-                .bytes()
-                .await
-                .expect("Could not read endpoint response body");
-            let body = bytes_to_string(bytes);
+            let attempt = match res.bytes().await {
+                Ok(bytes) => {
+                    let body = bytes_to_string(bytes);
 
-            let attempt = messageattempt::ActiveModel {
-                response_status_code: Set(status_code),
-                response: Set(body),
-                status: Set(status),
-                ..attempt
+                    messageattempt::ActiveModel {
+                        response_status_code: Set(status_code),
+                        response: Set(body),
+                        status: Set(status),
+                        ..attempt
+                    }
+                }
+
+                Err(err) => {
+                    tracing::warn!("Error reading response body: {}", err);
+                    messageattempt::ActiveModel {
+                        response_status_code: Set(status_code),
+                        response: Set(format!("failed to read response body: {}", err)),
+                        status: Set(status),
+                        ..attempt
+                    }
+                }
             };
+
             match http_error {
                 Some(err) => Err((attempt, err)),
                 None => Ok(attempt),
             }
         }
+
         Err(err) => {
             let attempt = messageattempt::ActiveModel {
                 response_status_code: Set(0),
@@ -406,21 +413,21 @@ async fn dispatch(
 
     match attempt {
         Ok(attempt) => {
-            let _attempt = attempt.insert(db).await?;
+            let _attempt = ctx!(attempt.insert(db).await)?;
 
             let msg_dest = messagedestination::ActiveModel {
                 status: Set(MessageStatus::Success),
                 next_attempt: Set(None),
                 ..msg_dest.into()
             };
-            let msg_dest = msg_dest.update(db).await?;
+            let msg_dest = ctx!(msg_dest.update(db).await)?;
 
             process_success_cache(cache, &msg_task.app_id, &msg_task.endpoint_id).await?;
 
             tracing::trace!("Worker success: {} {}", &msg_dest.id, &endp.id,);
         }
         Err((attempt, err)) => {
-            let attempt = attempt.insert(db).await?;
+            let attempt = ctx!(attempt.insert(db).await)?;
 
             let retry_schedule_override =
                 get_retry_schedule_override(db, cache, org_id, event_type_name)
@@ -458,7 +465,7 @@ async fn dispatch(
                     )),
                     ..msg_dest.into()
                 };
-                let _msg_dest = msg_dest.update(db).await?;
+                let _msg_dest = ctx!(msg_dest.update(db).await)?;
 
                 if attempt_count == OP_WEBHOOKS_SEND_FAILING_EVENT_AFTER {
                     op_webhook_sender
@@ -497,7 +504,7 @@ async fn dispatch(
                     next_attempt: Set(None),
                     ..msg_dest.into()
                 };
-                let _msg_dest = msg_dest.update(db).await?;
+                let _msg_dest = ctx!(msg_dest.update(db).await)?;
 
                 // Send common operational webhook
                 op_webhook_sender
@@ -541,17 +548,20 @@ async fn dispatch(
                             .await?;
 
                         // Disable endpoint in DB
-                        let endp = endpoint::Entity::secure_find_by_id(
-                            msg_task.app_id.clone(),
-                            msg_task.endpoint_id.clone(),
-                        )
-                        .one(db)
-                        .await?
+                        let endp = ctx!(
+                            endpoint::Entity::secure_find_by_id(
+                                msg_task.app_id.clone(),
+                                msg_task.endpoint_id.clone(),
+                            )
+                            .one(db)
+                            .await
+                        )?
                         .ok_or_else(|| {
-                            Error::Generic(format!(
+                            err_generic!(
                                 "Endpoint not found {} {}",
-                                &msg_task.app_id, &msg_task.endpoint_id
-                            ))
+                                &msg_task.app_id,
+                                &msg_task.endpoint_id
+                            )
                         })?;
 
                         let endp = endpoint::ActiveModel {
@@ -559,7 +569,7 @@ async fn dispatch(
                             first_failure_at: Set(Some(first_failure_at.into())),
                             ..endp.into()
                         };
-                        let _endp = endp.update(db).await?;
+                        let _endp = ctx!(endp.update(db).await)?;
                     }
                 }
             }
@@ -583,28 +593,26 @@ async fn get_retry_schedule_override(
 ) -> Result<Option<RetrySchedule>> {
     let cache_key = RetryScheduleCacheKey::new(org_id, event_type_name);
 
-    if let Some(retry_schedule) = cache
-        .get::<RetryScheduleCacheValue>(&cache_key)
-        .await
-        .map_err(|e| Error::from(e.to_string()))?
-    {
+    if let Some(retry_schedule) = ctx!(cache.get::<RetryScheduleCacheValue>(&cache_key).await)? {
         return Ok(retry_schedule.retry_schedule);
     };
 
-    let event_type =
+    let event_type = ctx!(
         eventtype::Entity::secure_find_by_name(org_id.to_owned(), event_type_name.to_owned())
             .one(db)
-            .await?;
+            .await
+    )?;
 
     match event_type {
         Some(event_type) => {
             let retry_schedule = event_type.retry_schedule;
             let cache_value = RetryScheduleCacheValue { retry_schedule };
 
-            cache
-                .set(&cache_key, &cache_value, Duration::from_secs(60))
-                .await
-                .map_err(|e| Error::from(e.to_string()))?;
+            ctx!(
+                cache
+                    .set(&cache_key, &cache_value, Duration::from_secs(60))
+                    .await
+            )?;
 
             Ok(cache_value.retry_schedule)
         }
@@ -636,10 +644,8 @@ async fn process_task(worker_context: WorkerContext<'_>, queue_task: Arc<QueueTa
         QueueTask::HealthCheck => unreachable!(),
     };
 
-    let msg = message::Entity::find_by_id(msg_id.clone())
-        .one(db)
-        .await?
-        .ok_or_else(|| Error::Generic(format!("Unexpected: message doesn't exist {}", msg_id,)))?;
+    let msg = ctx!(message::Entity::find_by_id(msg_id.clone()).one(db).await)?
+        .ok_or_else(|| err_generic!("Unexpected: message doesn't exist {}", msg_id,))?;
     let payload = msg.payload.as_ref().expect("Message payload is NULL");
 
     let create_message_app = CreateMessageApp::layered_fetch(
@@ -651,7 +657,7 @@ async fn process_task(worker_context: WorkerContext<'_>, queue_task: Arc<QueueTa
         Duration::from_secs(30),
     )
     .await?
-    .ok_or_else(|| Error::Generic(format!("Application doesn't exist: {}", &msg.app_id)))?;
+    .ok_or_else(|| err_generic!("Application doesn't exist: {}", &msg.app_id))?;
 
     let app_uid = create_message_app.uid.clone();
 
@@ -677,9 +683,11 @@ async fn process_task(worker_context: WorkerContext<'_>, queue_task: Arc<QueueTa
                 status: Set(MessageStatus::Sending),
                 ..Default::default()
             });
-        messagedestination::Entity::insert_many(destinations)
-            .exec(db)
-            .await?;
+        ctx!(
+            messagedestination::Entity::insert_many(destinations)
+                .exec(db)
+                .await
+        )?;
     }
 
     let org_id = &msg.org_id;
@@ -724,10 +732,10 @@ async fn process_task(worker_context: WorkerContext<'_>, queue_task: Arc<QueueTa
 
     let errs: Vec<_> = join.iter().filter(|x| x.is_err()).collect();
     if !errs.is_empty() {
-        return Err(Error::Generic(format!(
+        return Err(err_generic!(
             "Some dispatches failed unexpectedly: {:?}",
             errs
-        )));
+        ));
     }
 
     Ok(())